\name{mbatchWriteSuccessfulLog}
\alias{mbatchWriteSuccessfulLog}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{mbatchWriteSuccessfulLog}
\description{Internal use. Writes a known string to end of a log file.}
\usage{
mbatchWriteSuccessfulLog()
}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function ()
{
    logInfo("mbatchAssess Finishing")
  }
}
