\name{readAsGenericMatrix}
\alias{readAsGenericMatrix}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{readAsGenericMatrix}
\description{
Given a tab-delimited file with column and row headers, read it into a matrix.
}
\usage{
readAsGenericMatrix(theFile)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{theFile}{full path to file to load into matrix}
}
\details{
Reads a tab-delimited file using R in an efficient manner, into a matrix. First line should start with a tab.
}
\value{
a matrix based on the file read in
}
\references{}
\author{}
\note{}
\seealso{}
\examples{
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
